requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
werkzeug==3.0.3
scikit-learn==1.5.1
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
